tech(base): restructured the example app outside source
