+++ example-app/ErrorHandler/ErrorHandler.js
