+++ example-app/ErrorHandler/index.js
